# Conflicts:
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/db/models/Flight.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/db/models/FlightTicket.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/db/models/Passenger.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/db/models/Restaurant.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/db/models/RestaurantItem.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/db/models/Store.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/db/models/User.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/db/models/singletons/SingletonEnterprises.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/db/models/singletons/SingletonFlights.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/db/models/singletons/SingletonUser.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/network/ApiEndPoint.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/data/prefs/UserPreferences.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/AccountFragment.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/EditContactsFragment.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/EditPersonalDataFragment.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/LoginActivity.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/MainActivity.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/RestaurantDetailsActivity.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/RestaurantsFragment.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/StoreDetailsActivity.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/StoresFragment.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/SupportActivity.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/SupportTicketActivity.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/SupportTicketInfoActivity.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/TicketInfoActivity.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/TicketsActivity.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/utils/NetworkUtils.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/utils/UserJsonParser.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/utils/UserValidations.java
#	Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/utils/Validations.java
#	Aerocontrol/app/src/main/res/layout/activity_edit_account.xml
#	Aerocontrol/app/src/main/res/layout/activity_restaurant_details.xml
#	Aerocontrol/app/src/main/res/layout/activity_ticket_info.xml
#	Aerocontrol/app/src/main/res/layout/activity_tickets.xml
#	Aerocontrol/app/src/main/res/layout/gender_list_item.xml
#	Aerocontrol/app/src/main/res/layout/item_ticket_passengers.xml
#	Aerocontrol/app/src/main/res/values/strings.xml

+++ Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/MainActivity.java
@@ -54,14 +54,14 @@ public class MainActivity extends AppCompatActivity {
     public void showRestaurantDetails(final int id){
         Restaurant restaurant = SingletonEnterprises.getInstance(this).getRestaurantById(id);
         Intent intent = new Intent(this, RestaurantDetailsActivity.class);
-        intent.putExtra(RestaurantDetailsActivity.RESTAURANT_ID,(int) restaurant.getId());
+        intent.putExtra(RestaurantDetailsActivity.RESTAURANT_ID, restaurant.getId());
         startActivity(intent);
     }
 
     public void showStoreDetails(final int id){
         Store store = SingletonEnterprises.getInstance(this).getStoreById(id);
         Intent intent = new Intent(this, StoreDetailsActivity.class);
-        intent.putExtra(StoreDetailsActivity.STORE_ID, (int) store.getId());
+        intent.putExtra(StoreDetailsActivity.STORE_ID, store.getId());
         startActivity(intent);
     }
 }
